tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.pattern.basic.type.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.pattern.basic.type</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.pattern.basic.type</h1>
<p class="desc"></p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertDifferent(window.CanvasPattern, undefined, "window.CanvasPattern", "undefined");

  window.CanvasPattern.prototype.thisImplementsCanvasPattern = true;

  var response = await fetch('/images/green.png')
  var blob = await response.blob();
  var img = await createImageBitmap(blob);
  var pattern = ctx.createPattern(img, 'no-repeat');
  _assert(pattern.thisImplementsCanvasPattern, "pattern.thisImplementsCanvasPattern");

}, "");
</script>
